Fix warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@40763 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/tape/tap.c
@@ -1506,7 +1506,7 @@ int tap_seek_to_offset(tap_t *tap, unsigned long offset)
 {
     if (tap && tap->fd) {
         fseek(tap->fd, offset, SEEK_SET);
-        tap->current_file_seek_position = offset;
+        tap->current_file_seek_position = (int)offset;
         return 0;
     }
     return -1;
